jQuery Update
Added jQuery, fixed a ton of bugs and added remove button in project.
added "create loops" and "load loops" in app. fixed no modal problem, added "sweetAlert" so that you can do a more sweet alerts.
also did a lot of adding to the app. I updated the app now a lot, but the new version is coming. this is just the alpha.